+++ examples/webhooks-express-receiving.js
@@ -12,7 +12,7 @@ client.on('paid', (invoice, requestDate) => {
 
 app.use('/secret', client.middleware());
 
-// Important: if you want to use self-signed certificate
+// Note: if you want to use self-signed certificate
 // you must uploat it in CryptoBot API application settings
 createServer({
   key: readFileSync(__dirname + '/server.key'),