+++ src/classes/ClientEmitter.ts
@@ -88,7 +88,7 @@ class ClientEmitter extends Client {
   /**
    * Create handling webhooks server
    *
-   * Important: if you want to use self-signed certificate
+   * Note: if you want to use self-signed certificate
    * you must uploat it in CryptoBot API application settings
    *
    * @param serverOptions - Node.js built-in server options